Fix problem with disappearing thread notes
Every time a thread note was saved then the Workflow.Thread object
was replaced in the Workflow.Errand object with a new.
This caused thread notes to "disappear" (there really were in the old thread object).
Fix this by only creating a Workflow.Thread object if the errand does not have one.