Fix problem with HTML entities not displaying correctly
Some unsafe characters are converted to HTML entities.
The last character to be converted to a HTML entity was &.
That meant that all other HTML entities such as &gt; had their & converted to &amp;
which meant they didn't display the HTML entity they should have.

Solve this by converting & to an HTML entity last.